Fix dpr train
